requirements.txt
bottle
numpy
scipy
scikit-learn
matplotlib
jsonpickle
joblib
pyyaml
pandas